[USB] Disable USB pullup hack when internal pullups are available
Previously, when no external pullup control was available, the core
would apply a hack to override an external fixed pullup. However, on
chips that have internal pullups, these will be automatically
controlled, so this is not needed at all.

This probably needs to be more configurable by board, but this is enough
to for the specific board we are working with now.

See https://github.com/stm32duino/Arduino_Core_STM32/issues/885