[USB] Remove configuration and interface strings
These strings were defined and there was machinery to let the host
request them, but they were not actually referenced in any descriptors.
To clean up the code a bit, just remove the strings and the machinery.
Specifying the callback functions as NULL signals to the USBD core that
these strings are not available.