[USB] Set the bcdDevice value to 0
This value should probably be configurable by the sketch, so a
meaningful value can be picked. As long is this is not possible, best to
use 0, rather than 2.0, otherwise it later becomes impossible to use 2.0
in a meaningful way (you'll never know if the device is really 2.0, or
an old version that had no meaningful version number).