Clear reset flags before starting bootloader
Since the bootloader does not touch these flags, but can start the
sketch if instructed by the programmer (without a reset), clearing is
needed to prevent a start from the bootloader looking like a
software-reset. Otherwise, the startup code would be looking at the
BootIntoBootloaderAfterReset variable to decide whether to jump to the
bootloader. This variable should have been cleared before jumping to the
bootloader the last time, but the bootloader might have overwritten it
(or a sketch was uploaded with a different address for that variable).

Additionally, clearing the reset flags allows a sketch to see that the
they are started through the bootloader, when all reset flags are
cleared.