Jump to bootloader from asm block
This ensures that the SP is not modified after loading it and ensures a
jump instruction is used.

This also allows removing the volatile from sysMem_addr, since the
potential null-pointer dereference that results from it now happens in
assembly, so is no longer problematic.

The assembly code was based on micropython and another STM32 core:

    https://github.com/micropython/micropython/blob/f6375ac3ebac28656a0a757952d32c265b1ba7aa/ports/stm32/powerctrl.c#L71-L78
    https://github.com/GrumpyOldPizza/arduino-STM32L4/blob/bacc184288a4644b2ee6a97672334983f3e788ab/system/STM32L4xx/Source/boot_stm32l4xx.c#L159-L166