HACK: Do not use USB_DevDisconnect in early startup
This function calls HAL_Delay, which relies on the systick timer to be
running, so this results in an infinite loop. This should probably be
fixed in HAL_Delay, but for now just remove the calls to
USB_DevDisconnect and USB_DevConnect and replace it with the bit
twiddles needed for some chip families (e.g. the F4), breaking
compilation and/or functionality for others.